+++ ReleaseNotes.md
@@ -14,6 +14,7 @@
 - cta/CTA#1029 Fix segmentatin fault in frontend when list repacks of a tape that has been deleted in the catalogue
 - cta/CTA#1032 cta-admin dr ls crashes the frontend if executed during an archive/retrieve
 - cta/CTA#996  Removes PARALLEL from migration scripts
+- cta/CTA#1035 log configuration options on frontend startup
 
 # v4.2-2